requirements.txt
# CompactPDF - Dependências Essenciais
# Apenas PyMuPDF e Spire.PDF para compressão eficaz

# Método principal - PyMuPDF
pymupdf>=1.24.0

# Método alternativo - Spire.PDF
Spire.PDF>=9.0.0

# Dependências do sistema (geralmente já incluídas)
# tkinter - interface gráfica (incluído no Python)
# pathlib - manipulação de caminhos (incluído no Python 3.4+)
# threading - processamento paralelo (incluído no Python)
# enum - enumerações (incluído no Python 3.4+)
